Custom boilers

+++ heat-production-optimization/ProductionUnits.cs
@@ -23,7 +23,7 @@ namespace heat_production_optimization
 
     public class GasBoiler : ProductionUnitDataModel, IUnit
     {
-		public GasBoiler(Guid id, double maxHeat, int productionCost, int cO2Emission, double gasConsumption, string alias = "GB", string name = "Gas Boiler")
+		public GasBoiler(Guid id, double maxHeat, double productionCost, double cO2Emission, double gasConsumption, string alias = "GB", string name = "Gas Boiler")
         {
             Id = id;
 			if (maxHeat == 0) maxHeat = 0.01;
@@ -42,7 +42,7 @@ namespace heat_production_optimization
     public class OilBoiler : ProductionUnitDataModel, IUnit
     {
 
-		public OilBoiler(Guid id, double maxHeat, int productionCost, int cO2Emission, double oilConsumption, string alias = "OB", string name = "Oil Boiler")
+		public OilBoiler(Guid id, double maxHeat, double productionCost, double cO2Emission, double oilConsumption, string alias = "OB", string name = "Oil Boiler")
         {
             Id = id;
 			if (maxHeat == 0) maxHeat = 0.01;
@@ -61,7 +61,7 @@ namespace heat_production_optimization
     public class GasMotor : ProductionUnitDataModel, IUnit
     {
 
-		public GasMotor(Guid id, double maxHeat, double maxElectricity, int productionCost, int cO2Emission, double gasConsumption, string alias = "GM", string name = "Gas Motor")
+		public GasMotor(Guid id, double maxHeat, double maxElectricity, double productionCost, double cO2Emission, double gasConsumption, string alias = "GM", string name = "Gas Motor")
         {
             Id = id;
 			if (maxHeat == 0) maxHeat = 0.01;
@@ -81,7 +81,7 @@ namespace heat_production_optimization
     public class ElectricBoiler : ProductionUnitDataModel, IUnit
     {
 
-        public ElectricBoiler(Guid id, double maxHeat, double maxElectricity, int productionCost, int cO2Emission, string alias = "EK", string name = "Electric Boiler")
+        public ElectricBoiler(Guid id, double maxHeat, double maxElectricity, double productionCost, double cO2Emission, string alias = "EK", string name = "Electric Boiler")
         {
             Id = id;
 			if (maxHeat == 0) maxHeat = 0.01;
@@ -99,19 +99,8 @@ namespace heat_production_optimization
 
     public class CustomUnit : ProductionUnitDataModel, IUnit
     {
-		//public string Alias { get; }
-		//public string Name { get; }
-		//public double MaxHeat { get; }
-		//public double MaxElectricity { get; }
-		//public double ProductionCost { get; }
-		//public double ProductionCostMWh { get; }
-		//public double CO2Emission { get; }
-  //      public double CO2EmissionMWh { get; }
-		//public double GasConsumption { get; }
-		//public double OilConsumption { get; }
-		//public double PriceToHeatRatio { get; set; }
 
-		public CustomUnit(Guid id,string alias, string name, double maxHeat = 0, double maxElectricity = 0, int productionCost = 0, int cO2Emission = 0, double gasConsumption = 0, double oilConsumption = 0)
+		public CustomUnit(Guid id,string alias, string name, double maxHeat = 0, double maxElectricity = 0, double productionCost = 0, double cO2Emission = 0, double gasConsumption = 0, double oilConsumption = 0)
 		{
             Id = id;
             if(maxHeat == 0) maxHeat = 0.01;
